+++ aws/rust-runtime/aws-runtime/Cargo.toml
@@ -33,6 +33,7 @@ tracing = "0.1"
 uuid = { version = "1" }
 
 [dev-dependencies]
+arbitrary = "1.3"
 aws-credential-types = { path = "../aws-credential-types", features = ["test-util"] }
 aws-smithy-async = { path = "../../../rust-runtime/aws-smithy-async", features = ["test-util"] }
 aws-smithy-protocol-test = { path = "../../../rust-runtime/aws-smithy-protocol-test" }
