+++ aws/rust-runtime/aws-runtime/src/lib.rs
@@ -40,3 +40,9 @@ pub mod request_info;
 
 /// Interceptor that determines the clock skew between the client and service.
 pub mod service_clock_skew;
+
+/// Supporting code for extracting config from an AWS config file.
+pub mod profile;
+
+/// Filesystem utilities
+pub mod fs_util;
