+++ aws/rust-runtime/aws-runtime/src/profile.rs
@@ -1,0 +1,12 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+pub mod error;
+mod normalize;
+pub mod parse;
+pub mod profile_file;
+pub mod profile_set;
+pub mod section;
+pub mod source;
